+++ packages/google_maps_flutter/google_maps_flutter/CHANGELOG.md
@@ -1,7 +1,8 @@
-## NEXT
+## 2.0.7
 
 * Add iOS unit and UI integration test targets.
 * Exclude arm64 simulators in example app.
+* Remove references to the Android V1 embedding.
 
 ## 2.0.6