+++ packages/image_picker/image_picker_for_web/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.1.1
+
+* Implemented `getMultiImage`.
+* Initialized the following `XFile` attributes for picked files:
+  * `name`, `length`, `mimeType` and `lastModified`.
+
 # 2.1.0
 
 * Implemented `getImage`, `getVideo` and `getFile` methods that return `XFile` instances.